requirements.txt
--only-binary=:all:

Django>=4.2,<5.0
djangorestframework==3.15.0
psycopg2-binary==2.9.9
Pillow>=10.2.0
python-dotenv
djangorestframework-simplejwt
gunicorn
whitenoise
dj-database-url
django-allauth
cryptography
